directory refactor
updated version

added BaseFilterableTask

updated DataSetTask(BaseTask) to DataSetTask(BaseFilterableTask)
