+++ CloudHarvestCoreTasks/blueprints.py
@@ -5,9 +5,8 @@ This file provides the types of blueprints that can be registered by the CloudHa
 
 We make these distinctions because we don't want Agents to load API endpoints and vice versa.
 """
-
-from flask import Blueprint
 from CloudHarvestCorePluginManager.decorators import register_definition
+from flask import Blueprint
 from logging import getLogger
 
 logger = getLogger('harvest')
