+++ CloudHarvestCoreTasks/chains/__init__.py
@@ -1,0 +1,3 @@
+from CloudHarvestCoreTasks.chains.base import BaseTaskChain
+from CloudHarvestCoreTasks.chains.harvest import BaseHarvestTaskChain
+from CloudHarvestCoreTasks.chains.report import ReportTaskChain
