+++ CloudHarvestCoreTasks/factories.py
@@ -3,7 +3,9 @@ factories.py - This module contains functions for creating task chains from file
 """
 from logging import getLogger
 from typing import Any
-from CloudHarvestCoreTasks.base import BaseTaskChain, BaseTask
+
+from CloudHarvestCoreTasks.tasks.base import BaseTask
+from CloudHarvestCoreTasks.chains.base import BaseTaskChain
 
 logger = getLogger('harvest')
 
@@ -74,7 +76,7 @@ def task_chain_from_dict(template: dict, **kwargs) -> BaseTaskChain:
         task_chain_configuration = template[task_chain_registered_class_name]
 
     except IndexError:
-        from CloudHarvestCoreTasks.base import BaseHarvestException
+        from exceptions import BaseHarvestException
         raise BaseHarvestException('No task chain class found in the task chain configuration.')
 
     # Attempt to locate the identified class in the registry.
@@ -84,7 +86,7 @@ def task_chain_from_dict(template: dict, **kwargs) -> BaseTaskChain:
                                     name=task_chain_registered_class_name)[0]
 
     except IndexError:
-        from CloudHarvestCoreTasks.base import BaseHarvestException
+        from exceptions import BaseHarvestException
         raise BaseHarvestException(f'No task chain class found for {task_chain_registered_class_name}.')
 
     # Set the name of the task chain if it is not already set.
@@ -126,6 +128,10 @@ def task_from_dict(task_configuration: dict or BaseTask,
 
     # Instantiate the task with the templated configuration and return it
     class_configuration = templated_task_configuration.get(class_name) or {}
+
+    if isinstance(task_chain, BaseTaskChain):
+        class_configuration = {'filters': task_chain.filters} | class_configuration
+
     instantiated_class = task_class(task_chain=task_chain, **class_configuration)
 
     instantiated_class.original_template = task_configuration[class_name]
