+++ CloudHarvestCoreTasks/meta.json
@@ -7,5 +7,5 @@
     "license": "CC Attribution-NonCommercial-ShareAlike 4.0 International",
     "name": "CloudHarvestCoreTasks",
     "url": "https://github.com/Cloud-Harvest/CloudHarvestCoreTasks",
-    "version": "0.5.1"
+    "version": "0.6.0"
 }
