+++ CloudHarvestCoreTasks/tasks/__init__.py
@@ -1,0 +1,18 @@
+# Base Classes
+from CloudHarvestCoreTasks.tasks.base import (
+    BaseTask,
+    TaskStatusCodes
+)
+
+# Tasks
+from CloudHarvestCoreTasks.tasks.dataset import DataSetTask
+from CloudHarvestCoreTasks.tasks.dummy import DummyTask
+from CloudHarvestCoreTasks.tasks.error import ErrorTask
+from CloudHarvestCoreTasks.tasks.file import FileTask
+from CloudHarvestCoreTasks.tasks.harvest_update import HarvestUpdateTask
+from CloudHarvestCoreTasks.tasks.http import HttpTask
+from CloudHarvestCoreTasks.tasks.json import JsonTask
+from CloudHarvestCoreTasks.tasks.mongo import MongoTask
+from CloudHarvestCoreTasks.tasks.prune import PruneTask
+from CloudHarvestCoreTasks.tasks.redis import RedisTask
+from CloudHarvestCoreTasks.tasks.wait import WaitTask
