+++ CloudHarvestCoreTasks/tasks/prune.py
@@ -1,0 +1,46 @@
+from CloudHarvestCorePluginManager import register_definition
+from CloudHarvestCoreTasks.tasks.base import BaseTask, TaskStatusCodes
+
+
+@register_definition(name='prune', category='task')
+class PruneTask(BaseTask):
+    def __init__(self, previous_task_data: bool = False, stored_variables: bool = False, *args, **kwargs):
+        """
+        Prunes the task chain.
+
+        This method can be used to clear the data of previous tasks and/or the stored variables in the task chain.
+        This can be useful to free up memory during the execution of a long task chain.
+
+        Args:
+            previous_task_data (bool, optional): If True, the data of all previous tasks in the task chain will be cleared. Defaults to False.
+            stored_variables (bool, optional): If True, all variables stored in the task chain will be cleared. Defaults to False.
+
+        Returns:
+            BaseTaskChain: The current instance of the task chain.
+        """
+
+        super().__init__(*args, **kwargs)
+        self.previous_task_data = previous_task_data
+        self.stored_variables = stored_variables
+
+    def method(self) -> 'PruneTask':
+        from sys import getsizeof
+        total_bytes_pruned = 0
+
+        # If previous_task_data is True, clear the data of all previous tasks
+        if self.previous_task_data:
+            for i in range(self.task_chain.position):
+                if str(self.task_chain[i].status) in [TaskStatusCodes.complete, TaskStatusCodes.error, TaskStatusCodes.skipped]:
+                    total_bytes_pruned += getsizeof(self.task_chain[i].result)
+                    self.task_chain[i].result = None
+
+        # If stored_variables is True, clear all variables stored in the task chain
+        if self.stored_variables:
+            total_bytes_pruned += getsizeof(self.task_chain.variables)
+            self.task_chain.variables.clear()
+
+        self.result = {
+            'total_bytes_pruned': total_bytes_pruned
+        }
+
+        return self
